backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
eventlet==0.33.3
python-socketio==5.9.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
